Moving some CharacterApi functions to appropriate utility classes; others to their respective domain.
