+++ src/interp/commands/Info.py
@@ -5,6 +5,7 @@ from injector import inject
 from game.EnumProvider import EnumProvider
 from game.RegistryService import RegistryService
 from api.ItemApi import ItemApi
+from util.EffectUtil import EffectUtil
 from util.GenericUtil import GenericUtil
 from game.WeatherHandler import WeatherHandler
 from api.InterpApi import InterpApi
@@ -86,7 +87,7 @@ class Info:
     def do_who(self, character: Character) -> str:
         who_list = [character] + PlayerUtil.visible(character, self.session_handler)
         lines = [
-            f"{CharacterApi.who_line(character, c)}\r\n"
+            f"{InfoUtil.who_line(character, c)}\r\n"
             for c in who_list
         ]
         lines.append(f"Players found: {len(who_list)}\r\n")
@@ -156,7 +157,7 @@ class Info:
             desc = "You see nothing special."
 
         lines = [desc, InfoUtil.target_condition_line(target)]
-        equip_lines = CharacterApi.target_equipment_lines(target, EQUIP_SLOT_LABELS)
+        equip_lines = InfoUtil.target_equipment_lines(target, EQUIP_SLOT_LABELS)
         if equip_lines:
             lines.append("")
             lines.append(f"{(target.name or 'They')} is using:")
@@ -314,7 +315,7 @@ class Info:
         if hunger == 0:
             lines.append("You are hungry.")
 
-        position_line = CharacterApi.score_position_line(attributes)
+        position_line = InfoUtil.score_position_line(attributes)
         lines.append(position_line)
 
         ac_pierce = character.armor_class.get_ac(character, 0)
@@ -350,7 +351,7 @@ class Info:
         lines.append(f"You are {InfoUtil.score_alignment_word(alignment)}.")
         if CharacterApi.is_comm_enabled(character, "COMM_SHOW_AFFECTS"):
             lines.append("")
-            lines.append(CharacterApi.format_affects(character).rstrip("\r\n"))
+            lines.append(EffectUtil.format_affects(character).rstrip("\r\n"))
         context.finish()
         return "\r\n".join(lines) + "\r\n"
 
@@ -588,7 +589,7 @@ class Info:
 
     def do_affects(self, character: Character, context: Context) -> str:
         context.finish()
-        return CharacterApi.format_affects(character)
+        return EffectUtil.format_affects(character)
 
     def do_autolist(self, character: Character, context: Context) -> str:
         if CharacterApi.is_npc(character):
@@ -773,7 +774,7 @@ class Info:
             return payload.get("to_char", "")
 
         lines = [
-            f"{CharacterApi.who_line(character, c)}\r\n"
+            f"{InfoUtil.who_line(character, c)}\r\n"
             for c in matches
         ]
         return "".join(lines)
@@ -984,7 +985,7 @@ class Info:
         return self._render_message_key(context, "set", channel="to_char", s=shown).get("to_char", "")
 
     def do_equipment(self, character: Character, context: Context) -> str:
-        lines = CharacterApi.target_equipment_lines(character, EQUIP_SLOT_LABELS)
+        lines = InfoUtil.target_equipment_lines(character, EQUIP_SLOT_LABELS)
         context.finish()
         if not lines:
             return "You are using:\r\nNothing.\r\n"
@@ -998,8 +999,8 @@ class Info:
         else:
             arg1 = (context.parameters[0] if context.parameters and len(context.parameters) > 0 else "").strip().lower()
             arg2 = (context.parameters[1] if context.parameters and len(context.parameters) > 1 else "").strip().lower()
-        obj1 = CharacterApi.find_owned_item(character, arg1) if arg1 else None
-        obj2 = CharacterApi.find_owned_item(character, arg2) if arg2 else (ItemApi.find_comparable_equipped_item(character, obj1) if obj1 is not None else None)
+        obj1 = character.find_owned_item(arg1) if arg1 else None
+        obj2 = character.find_owned_item(arg2) if arg2 else (ItemApi.find_comparable_equipped_item(character, obj1) if obj1 is not None else None)
         t1 = str(getattr(obj1, "item_type", "") or "").strip().lower() if obj1 is not None else ""
         t2 = str(getattr(obj2, "item_type", "") or "").strip().lower() if obj2 is not None else ""
         v1 = ItemApi.compare_value(obj1) if obj1 is not None else None
